+++ src/lib.rs
@@ -1,3 +1,3 @@
-mod ast;
-mod tokenizer;
 mod parser;
+mod tokenizer;
+mod types;
